Edit data visualization quantity

+++ IFarm/src/ifarm/DataEntryHandler.java
@@ -40,7 +40,7 @@ public class DataEntryHandler  {
                 String joined =  "User-" + s[0] + " Farm-" +  s[1] + " " +s[4] + " " + s[5] + " " +  s[3] + " " + s[6] + "" + s[7] + " " + s[8] + " " + s[9] + " " + atomicInteger.getAndIncrement() ;
                 System.out.println("DATA ENTRY: "+Thread.currentThread().getName() + ": " + joined);
                 Date date = Date.valueOf(s[4]);
-                activityClass.toDB(atomicInteger.get(), date, s[5], s[3], s[7], Integer.parseInt(s[6]), Integer.parseInt(s[8]), Integer.parseInt(s[9]), Integer.parseInt(s[1]), Integer.parseInt(s[0]));
+                activityClass.toDB(atomicInteger.get(), date, s[5], s[3], s[7], Float.parseFloat(s[6]), Integer.parseInt(s[8]), Integer.parseInt(s[9]), Integer.parseInt(s[1]), Integer.parseInt(s[0]));
                 activityClass.toTxt(joined, s[0]);
             }
         } finally {
